+++ context/app/static/js/pages/CellsAPIDemo/ResultsTable.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableRow from '@material-ui/core/TableRow';
+import TableCell from '@material-ui/core/TableCell';
+import TableHead from '@material-ui/core/TableHead';
+
+function ResultsTable(props) {
+  const { results } = props;
+  if (results.length === 0) {
+    return <p>No results</p>;
+  }
+  const fields = Object.keys(results[0]);
+  return (
+    <Table>
+      <TableHead>
+        <TableRow>
+          {fields.map((field) => (
+            <TableCell key={field}>{field}</TableCell>
+          ))}
+        </TableRow>
+      </TableHead>
+      <TableBody>
+        {results.map((result) => (
+          <TableRow>
+            {fields.map((field) => (
+              <TableCell key={field}>{result[field]}</TableCell>
+            ))}
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  );
+}
+
+export default ResultsTable;
